split Navbar into multiple components
